+++ src/OpenTelemetry/Metrics/MetricType.cs
@@ -27,7 +27,7 @@ namespace OpenTelemetry.Metrics
             0x20: Gauge
             0x30: Summary (reserved)
             0x40: Histogram
-            0x50: HistogramWithMinMax
+            0x50: HistogramWithMinMax (reserved)
             0x60: ExponentialHistogram (reserved)
             0x70: ExponentialHistogramWithMinMax (reserved)
             0x80: SumNonMonotonic
@@ -69,11 +69,6 @@ namespace OpenTelemetry.Metrics
         /// Histogram.
         /// </summary>
         Histogram = 0x40,
-
-        /// <summary>
-        /// Histogram with minimum and maximum values.
-        /// </summary>
-        HistogramWithMinMax = 0x50,
 
         /// <summary>
         /// Non-monotonic Sum of Long type.